Extract more reducer helpers
